refactor: create a cozy-helper to init cozy client
