+++ mobile/src/actions/index.js
@@ -1,7 +1,7 @@
 /* global __ALLOW_HTTP__ */
 
+import { init } from '../lib/cozy-helper'
 import cozy from 'cozy-client-js'
-import { init } from '../lib/cozy_init'
 import localforage from 'localforage'
 
 import { DISPLAY_TOAST } from '../../../src/actions'
@@ -26,77 +26,71 @@ export class OnBoardingError extends Error {
   }
 }
 
-export const setUrl = (url) => {
-  return async dispatch => {
-    let scheme = 'https://'
-    if (__ALLOW_HTTP__) {
-      scheme = 'http://'
-      console.warn('development mode: we don\'t check SSL requirement')
+export const checkURL = url => async dispatch => {
+  let scheme = 'https://'
+  if (__ALLOW_HTTP__) {
+    scheme = 'http://'
+    console.warn('development mode: we don\'t check SSL requirement')
+  }
+  if (/(.*):\/\/(.*)/.test(url) && !url.startsWith(scheme)) {
+    dispatch(wrongAddressError())
+    throw new OnBoardingError(`The only supported protocol is ${scheme}`)
+  }
+  if (!url.startsWith(scheme)) {
+    url = `${scheme}${url}`
+  }
+  return dispatch({ type: SET_URL, url: url })
+}
+
+const openRegistrationWith = inAppBrowser => new Promise((resolve) => {
+  inAppBrowser.addEventListener('loadstart', ({url}) => {
+    const accessCode = /\?access_code=(.+)$/.test(url)
+    const state = /\?state=(.+)$/.test(url)
+
+    if (accessCode || state) {
+      resolve(url)
     }
-    if (/(.*):\/\/(.*)/.test(url) && !url.startsWith(scheme)) {
-      dispatch(wrongAddressError())
-      throw new OnBoardingError(`The only supported protocol is ${scheme}`)
-    }
-    if (!url.startsWith(scheme)) {
-      url = `${scheme}${url}`
-    }
-    return dispatch({ type: SET_URL, url: url })
+  })
+})
+
+const onRegistered = dispatch => (client, url) => {
+  if (window.cordova && window.cordova.InAppBrowser) {
+    const target = '_blank'
+    const options = 'location=yes,hidden=no'
+    const inAppBrowser = window.cordova.InAppBrowser.open(url, target, options)
+    return openRegistrationWith(inAppBrowser)
+    .then(
+      url => {
+        inAppBrowser.close()
+        return url
+      },
+      err => {
+        inAppBrowser.close()
+        dispatch(wrongAddressError())
+        throw err
+      }
+    )
   }
 }
 
-const onRegistered = (dispatch) => {
-  const { cordova } = window
-  return (client, url) => {
-    if (cordova && cordova.InAppBrowser) {
-      const { InAppBrowser } = cordova
-      const target = '_blank'
-      const options = 'location=yes,hidden=no'
-      const inAppBrowser = InAppBrowser.open(url, target, options)
-
-      return new Promise((resolve) => {
-        inAppBrowser.addEventListener('loadstart', ({url}) => {
-          const accessCode = /\?access_code=(.+)$/.test(url)
-          const state = /\?state=(.+)$/.test(url)
-
-          if (accessCode || state) {
-            resolve(url)
-          }
-        })
-      })
-      .then(
-        (url) => {
-          inAppBrowser.close()
-          return url
-        },
-        (err) => {
-          inAppBrowser.close()
-          dispatch(error())
-          throw err
-        }
-      )
-    }
+export const registerDevice = (router, location) => async (dispatch, getState) => {
+  await dispatch(checkURL(getState().mobile.serverUrl))
+  const device = window.cordova ? window.cordova.platformId : null
+  await init(getState().mobile.serverUrl, onRegistered(dispatch), device)
+  try {
+    await cozy.authorize()
+    await cozy.offline.replicateFromCozy('io.cozy.files')
+  } catch (err) {
+    dispatch(wrongAddressError())
+    throw err
   }
-}
 
-// TODO need to refactor this braces hell
-export const registerDevice = (router, location) => {
-  return async (dispatch, getState) => {
-    await dispatch(setUrl(getState().mobile.serverUrl))
-    await init(getState().mobile.serverUrl, onRegistered(dispatch))
-    try {
-      await cozy.authorize()
-      await cozy.offline.replicateFromCozy('io.cozy.files')
-    } catch (err) {
-      dispatch(wrongAddressError())
-      throw err
-    }
-
-    dispatch({ type: SETUP })
-    localforage.setItem('state', getState().mobile)
-    if (location.state && location.state.nextPathname) {
-      router.replace(location.state.nextPathname)
-    } else {
-      router.replace('/')
-    }
+  // TODO move this outside of this action (may be in the smart component's behavior)
+  dispatch({ type: SETUP })
+  localforage.setItem('state', getState().mobile)
+  if (location.state && location.state.nextPathname) {
+    router.replace(location.state.nextPathname)
+  } else {
+    router.replace('/')
   }
 }
